fix(op): 修复 site 测试时间戳字面量在 Go 1.26 下的溢出

+++ internal/op/site_test.go
@@ -1140,7 +1140,7 @@ func buildAllAPIHubImportPayload(managedUsername string) map[string]any {
 func buildMetAPIImportPayload(managedUsername string) map[string]any {
 	return map[string]any{
 		"version":   "2.1",
-		"timestamp": 1760000000000,
+		"timestamp": int64(1760000000000),
 		"type":      "accounts",
 		"accounts": map[string]any{
 			"sites": []any{
